Started working on the beginnings of message parsing.
Basic mbox and message validation has been implemented.  It extracts the
sender address from the current mbox message.  However, that's about as
far as I've gotten for now.

Created a Message type, but it's increasingly looking like a useless
type.  Will be removed in the next commit.  For now, though, everything
is working.