mbox now parses headers to my satisfaction.
The Mbox library now parses headers properly.  I said that I was going to
remove the Message type before, but I'm going to keep it for now, more for
developmental convenience of not having to refactor the interface to mbox just
yet.  However, the interface isn't all that bad so far.  I might end up keeping
it the way it is.