Implemented body reader.
Research shows that the io.Reader interface is the traditional means of
reading the contents of structured data.  This makes some sense, as it
allows the use of fixed-sized buffers, which is what distinguishes Go
from other languages in high-performance applications.  While mbox is
hardly intended for use with high-performance applications, it's best to
stick with established community conventions when building a Go API.