Removed debugging output, wrapping up API
The mbox API seems to now be functionality-complete.  It now handles
multiple messages in an MBOX stream, it delivers an EOF when it reaches
the end of the stream, and it seems to properly frame message body and
headers alike.  It lacks some convenience functions, but those can come
later.  For now, my next task is documentation.